services/admin-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9

